+++ apps/terminal/api/v1/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+#
+from .terminal import *
+from .session import *
+from .task import *
