+++ apps/terminal/api/v2/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+#
+from .terminal import *
