+++ apps/users/api/v2/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+#
+
+from .user import *
