+++ apps/users/api/v2/user.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+#
+from rest_framework import viewsets
+
+from common.permissions import WithBootstrapToken
+from ...serializers import v2 as serializers
+
+
+class ServiceAccountRegistrationViewSet(viewsets.ModelViewSet):
+    serializer_class = serializers.ServiceAccountRegistrationSerializer
+    permission_classes = (WithBootstrapToken,)
+    http_method_names = ['post']
