+++ apps/users/serializers/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+#
+from .v1 import *
