+++ apps/users/urls/api_urls.py
@@ -15,8 +15,7 @@ router.register(r'groups', api.UserGroupViewSet, 'user-group')
 
 
 urlpatterns = [
-    # path(r'', api.UserListView.as_view()),
-    path('token/', api.UserToken.as_view(), name='user-token'),
+    # path('token/', api.UserToken.as_view(), name='user-token'),
     path('connection-token/', api.UserConnectionTokenApi.as_view(), name='connection-token'),
     path('profile/', api.UserProfileApi.as_view(), name='user-profile'),
     path('auth/', api.UserAuthApi.as_view(), name='user-auth'),
@@ -31,5 +30,6 @@ urlpatterns = [
     path('users/<uuid:pk>/groups/', api.UserUpdateGroupApi.as_view(), name='user-update-group'),
     path('groups/<uuid:pk>/users/', api.UserGroupUpdateUserApi.as_view(), name='user-group-update-user'),
 ]
-
 urlpatterns += router.urls
+
+
