+++ utils/create_assets_user/admin_users.txt
@@ -1,0 +1,3 @@
+# 名称 用户名 密码
+test123 testq12 test123123123
+
